+++ tesseract_command_language/src/cartesian_waypoint.cpp
@@ -83,5 +83,5 @@ void tesseract_planning::CartesianWaypoint::serialize(Archive& ar, const unsigne
 
 #include <tesseract_common/serialization.h>
 
-TESSERACT_CARTESIAN_WAYPOINT_EXPORT_IMPLEMENT(tesseract_planning::CartesianWaypoint)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::CartesianWaypoint)
+TESSERACT_CARTESIAN_WAYPOINT_EXPORT_IMPLEMENT(tesseract_planning::CartesianWaypoint)