+++ tesseract_examples/test/glass_upright_example_unit.cpp
@@ -6,7 +6,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_examples/glass_upright_example.h>
 #include <tesseract_common/filesystem.h>
 #include <tesseract_environment/environment.h>
-#include <tesseract_support/tesseract_support_resource_locator.h>
+#include <tesseract_common/resource_locator.h>
 
 using namespace tesseract_examples;
 using namespace tesseract_common;
@@ -14,7 +14,7 @@ using namespace tesseract_environment;
 
 TEST(TesseractExamples, GlassUprightTrajOptExampleUnit)  // NOLINT
 {
-  auto locator = std::make_shared<TesseractSupportResourceLocator>();
+  auto locator = std::make_shared<GeneralResourceLocator>();
   tesseract_common::fs::path urdf_path =
       locator->locateResource("package://tesseract_support/urdf/lbr_iiwa_14_r820.urdf")->getFilePath();
   tesseract_common::fs::path srdf_path =
@@ -29,7 +29,7 @@ TEST(TesseractExamples, GlassUprightTrajOptExampleUnit)  // NOLINT
 
 TEST(TesseractExamples, GlassUprightTrajOptIfoptExampleUnit)  // NOLINT
 {
-  auto locator = std::make_shared<TesseractSupportResourceLocator>();
+  auto locator = std::make_shared<GeneralResourceLocator>();
   tesseract_common::fs::path urdf_path =
       locator->locateResource("package://tesseract_support/urdf/lbr_iiwa_14_r820.urdf")->getFilePath();
   tesseract_common::fs::path srdf_path =