+++ tesseract_motion_planners/examples/freespace_example.cpp
@@ -50,7 +50,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_command_language/cartesian_waypoint.h>
 #include <tesseract_command_language/move_instruction.h>
 #include <tesseract_command_language/utils.h>
-#include <tesseract_support/tesseract_support_resource_locator.h>
+#include <tesseract_common/resource_locator.h>
 
 using namespace tesseract_planning;
 
@@ -62,10 +62,12 @@ int main(int /*argc*/, char** /*argv*/)
   try
   {
     // Setup
-    auto locator = std::make_shared<tesseract_common::TesseractSupportResourceLocator>();
+    auto locator = std::make_shared<tesseract_common::GeneralResourceLocator>();
     auto env = std::make_shared<tesseract_environment::Environment>();
-    tesseract_common::fs::path urdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/abb_irb2400.urdf");
-    tesseract_common::fs::path srdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/abb_irb2400.srdf");
+    tesseract_common::fs::path urdf_path(
+        locator->locateResource("package://tesseract_support/urdf/abb_irb2400.urdf")->getFilePath());
+    tesseract_common::fs::path srdf_path(
+        locator->locateResource("package://tesseract_support/urdf/abb_irb2400.srdf")->getFilePath());
     env->init(urdf_path, srdf_path, locator);
 
     // Dynamically load ignition visualizer if exist