+++ tesseract_motion_planners/trajopt/test/trajopt_planner_tests.cpp
@@ -52,7 +52,7 @@ TESSERACT_COMMON_IGNORE_WARNINGS_POP
 #include <tesseract_motion_planners/core/types.h>
 #include <tesseract_motion_planners/core/utils.h>
 
-#include <tesseract_support/tesseract_support_resource_locator.h>
+#include <tesseract_common/resource_locator.h>
 
 const int NUM_STEPS = 7;
 
@@ -72,10 +72,12 @@ protected:
 
   void SetUp() override
   {
-    auto locator = std::make_shared<tesseract_common::TesseractSupportResourceLocator>();
+    auto locator = std::make_shared<tesseract_common::GeneralResourceLocator>();
     Environment::Ptr env = std::make_shared<Environment>();
-    tesseract_common::fs::path urdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/lbr_iiwa_14_r820.urdf");
-    tesseract_common::fs::path srdf_path(std::string(TESSERACT_SUPPORT_DIR) + "/urdf/lbr_iiwa_14_r820.srdf");
+    tesseract_common::fs::path urdf_path(
+        locator->locateResource("package://tesseract_support/urdf/lbr_iiwa_14_r820.urdf")->getFilePath());
+    tesseract_common::fs::path srdf_path(
+        locator->locateResource("package://tesseract_support/urdf/lbr_iiwa_14_r820.srdf")->getFilePath());
     EXPECT_TRUE(env->init(urdf_path, srdf_path, locator));
     env_ = env;
     manip.tcp_frame = "tool0";