+++ tesseract_task_composer/core/src/task_composer_graph.cpp
@@ -568,5 +568,5 @@ void TaskComposerGraph::serialize(Archive& ar, const unsigned int /*version*/)
 
 }  // namespace tesseract_planning
 
-BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::TaskComposerGraph)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::TaskComposerGraph)
+BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::TaskComposerGraph)