+++ tesseract_task_composer/core/src/task_composer_node.cpp
@@ -618,5 +618,5 @@ void TaskComposerNode::setData(TaskComposerDataStorage& data_storage,
 
 }  // namespace tesseract_planning
 
-BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::TaskComposerNode)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::TaskComposerNode)
+BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::TaskComposerNode)