+++ tesseract_task_composer/core/src/task_composer_node_ports.cpp
@@ -153,5 +153,5 @@ void TaskComposerNodePorts::serialize(Archive& ar, const unsigned int /*version*
 }
 
 }  // namespace tesseract_planning
-BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::TaskComposerNodePorts)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::TaskComposerNodePorts)
+BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::TaskComposerNodePorts)