+++ tesseract_task_composer/core/src/test_suite/test_task.cpp
@@ -128,5 +128,5 @@ std::unique_ptr<TaskComposerNodeInfo> TestTask::runImpl(TaskComposerContext& con
 
 }  // namespace tesseract_planning::test_suite
 
-BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::test_suite::TestTask)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::test_suite::TestTask)
+BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::test_suite::TestTask)