+++ tesseract_task_composer/planning/src/nodes/discrete_contact_check_task.cpp
@@ -197,5 +197,5 @@ void DiscreteContactCheckTask::serialize(Archive& ar, const unsigned int /*versi
 
 }  // namespace tesseract_planning
 
-BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::DiscreteContactCheckTask)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::DiscreteContactCheckTask)
+BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::DiscreteContactCheckTask)