+++ tesseract_task_composer/planning/src/nodes/fix_state_bounds_task.cpp
@@ -286,5 +286,5 @@ void FixStateBoundsTask::serialize(Archive& ar, const unsigned int /*version*/)
 
 }  // namespace tesseract_planning
 
-BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::FixStateBoundsTask)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::FixStateBoundsTask)
+BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::FixStateBoundsTask)