+++ tesseract_task_composer/planning/src/nodes/update_start_state_task.cpp
@@ -153,5 +153,5 @@ void UpdateStartStateTask::serialize(Archive& ar, const unsigned int /*version*/
 
 }  // namespace tesseract_planning
 
-BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::UpdateStartStateTask)
 TESSERACT_SERIALIZE_ARCHIVES_INSTANTIATE(tesseract_planning::UpdateStartStateTask)
+BOOST_CLASS_EXPORT_IMPLEMENT(tesseract_planning::UpdateStartStateTask)