+++ tesseract_motion_planners/CHANGELOG.rst
@@ -278,8 +278,8 @@ Changelog for package tesseract_motion_planners
 
 0.9.5 (2022-03-31)
 ------------------
-* Update to leverage TesseractSupportResourceLocator (`#181 <https://github.com/tesseract-robotics/tesseract_planning/issues/181>`_)
-  * Update to leverage TesseractSupportResourceLocator
+* Update to leverage GeneralResourceLocator (`#181 <https://github.com/tesseract-robotics/tesseract_planning/issues/181>`_)
+  * Update to leverage GeneralResourceLocator
   * Update CI docker tag to 0.9
 * Contributors: Levi Armstrong